embedding_containers/requirements-lock.txt
certifi==2018.8.24
inflection==0.3.1
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.13
wincertstore==0.2
xlrd==1.1.0
